Renamed Evidence to TransporterCEG and moved removed from CEG class.
