+++ src/cegpy/graphs/__init__.py
@@ -1,1 +1,2 @@
 from .ceg import ChainEventGraph
+from .tceg import TransporterChainEventGraph
